requirements/common.txt
flask
gunicorn
dotenv
sklearn
nltk
newsapi-python
